Tests/LibWeb/Text/input/WebAudio/GainNode.html

<script src="../include.js"></script>
<script>
    function dumpAudioParam(param) {
        println(`${param} current: ${param.value}, default: ${param.defaultValue}, min: ${param.minValue}, max: ${param.maxValue}, rate: ${param.automationRate}`);
    }

    test(() => {
        const audioContext = new OfflineAudioContext(1, 5000, 44100);

        const node = audioContext.createGain();

        // Check prototype
        let prototype = Object.getPrototypeOf(node);

        while (prototype) {
            println(prototype.constructor.name);
            prototype = Object.getPrototypeOf(prototype);
        }

        // Audio Params
        dumpAudioParam(node.gain);
    });
</script>
